Use std::optional whenever possible
