+++ libs/core/datastructures/include/hpx/datastructures/optional.hpp
@@ -1,4 +1,4 @@
-//  Copyright (c) 2017-2022 Hartmut Kaiser
+//  Copyright (c) 2022 Hartmut Kaiser
 //
 //  SPDX-License-Identifier: BSL-1.0
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
@@ -8,701 +8,33 @@
 
 #include <hpx/config.hpp>
 
-#include <cstddef>
-#include <exception>
-#include <functional>
-#include <new>
-#include <stdexcept>
-#include <string>
-#include <type_traits>
-#include <utility>
+#if defined(HPX_HAVE_CXX17_COPY_ELISION)
+#include <functional>    // for std::hash<optional<T>>
+#include <optional>
 
 namespace hpx {
 
-    struct nullopt_t
-    {
-        struct init
-        {
-        };
-        constexpr explicit nullopt_t(nullopt_t::init) noexcept {}
-    };
-    constexpr nullopt_t nullopt{nullopt_t::init()};
-
-    class bad_optional_access : public std::logic_error
-    {
-    public:
-        explicit bad_optional_access(std::string const& what_arg)
-          : std::logic_error(what_arg)
-        {
-        }
-
-        explicit bad_optional_access(char const* what_arg)
-          : std::logic_error(what_arg)
-        {
-        }
-    };
-
-    namespace _optional_swap {
-
-        using std::swap;
-
-        template <typename T>
-        void check_swap() noexcept(
-            noexcept(swap(std::declval<T&>(), std::declval<T&>())));
-    }    // namespace _optional_swap
-
-    template <typename T>
-    class optional
-    {
-    public:
-        using value_type = T;
-
-        constexpr optional() noexcept
-          : empty_(true)
-        {
-        }
-
-        constexpr optional(nullopt_t) noexcept
-          : empty_(true)
-        {
-        }
-
-        optional(optional const& other)
-          : empty_(true)
-        {
-            if (!other.empty_)
-            {
-                new (&storage_) T(other.value());
-                empty_ = false;
-            }
-        }
-        optional(optional&& other) noexcept(
-            std::is_nothrow_move_constructible_v<T>)
-          : empty_(true)
-        {
-            if (!other.empty_)
-            {
-                new (&storage_) T(HPX_MOVE(other.value()));
-                empty_ = false;
-            }
-        }
-
-        template <typename U = T,
-            typename = std::enable_if_t<std::is_constructible_v<T, U&&>>>
-        explicit optional(U&& val) noexcept(
-            std::is_nothrow_move_constructible_v<T>)
-          : empty_(true)
-        {
-            new (&storage_) T(HPX_MOVE(val));
-            empty_ = false;
-        }
-
-        template <typename U,
-            typename = std::enable_if_t<std::is_constructible_v<T, U const&>>>
-        constexpr optional(optional<U> const& val)
-          : empty_(true)
-        {
-            if (val.has_value())
-            {
-                new (&storage_) T(*val);
-                empty_ = false;
-            }
-        }
-        template <typename U,
-            typename = std::enable_if_t<std::is_constructible_v<T, U&&>>>
-        constexpr optional(optional<U>&& val) noexcept(
-            std::is_nothrow_move_constructible_v<T>)
-          : empty_(true)
-        {
-            if (val.has_value())
-            {
-                new (&storage_) T(HPX_MOVE(*val));
-                empty_ = false;
-            }
-        }
-
-        template <typename... Ts>
-        explicit optional(std::in_place_t, Ts&&... ts)
-          : empty_(true)
-        {
-            new (&storage_) T(HPX_FORWARD(Ts, ts)...);
-            empty_ = false;
-        }
-
-        template <typename U, typename... Ts>
-        explicit optional(
-            std::in_place_t, std::initializer_list<U> il, Ts&&... ts)
-          : empty_(true)
-        {
-            new (&storage_) T(il, HPX_FORWARD(Ts, ts)...);
-            empty_ = false;
-        }
-
-        ~optional()
-        {
-            reset();
-        }
-
-        optional& operator=(optional const& other)
-        {
-            optional tmp(other);
-            swap(tmp);
-
-            return *this;
-        }
-        optional& operator=(optional&& other) noexcept(
-            std::is_nothrow_move_assignable_v<T>&&
-                std::is_nothrow_move_constructible_v<T>)
-        {
-            if (this == &other)
-            {
-                return *this;
-            }
-
-            if (empty_)
-            {
-                if (!other.empty_)
-                {
-                    new (&storage_) T(HPX_MOVE(other.value()));
-                    empty_ = false;
-                }
-            }
-            else
-            {
-                if (other.empty_)
-                {
-                    reinterpret_cast<T*>(&storage_)->~T();
-                    empty_ = true;
-                }
-                else
-                {
-                    **this = HPX_MOVE(other.value());
-                }
-            }
-            return *this;
-        }
-
-        optional& operator=(T&& other)
-        {
-            if (!empty_)
-            {
-                reinterpret_cast<T*>(&storage_)->~T();
-                empty_ = true;
-            }
-
-            new (&storage_) T(HPX_MOVE(other));
-            empty_ = false;
-
-            return *this;
-        }
-        template <typename U,
-            typename Enable =
-                std::enable_if_t<std::is_constructible_v<T, U const&>>>
-        optional& operator=(optional<U> const& other)
-        {
-            optional tmp(other);
-            swap(tmp);
-
-            return *this;
-        }
-        template <typename U,
-            typename Enable = std::enable_if_t<std::is_constructible_v<T, U&&>>>
-        optional& operator=(optional<U>&& other)
-        {
-            optional tmp(HPX_MOVE(other));
-            swap(tmp);
-
-            return *this;
-        }
-
-        optional& operator=(nullopt_t) noexcept
-        {
-            if (!empty_)
-            {
-                reinterpret_cast<T*>(&storage_)->~T();
-                empty_ = true;
-            }
-            return *this;
-        }
-
-        ///////////////////////////////////////////////////////////////////////
-        constexpr T const* operator->() const noexcept
-        {
-            return reinterpret_cast<T const*>(&storage_);
-        }
-
-        T* operator->() noexcept
-        {
-            return reinterpret_cast<T*>(&storage_);
-        }
-
-        constexpr T const& operator*() const& noexcept
-        {
-            return *reinterpret_cast<T const*>(&storage_);
-        }
-        T& operator*() & noexcept
-        {
-            return *reinterpret_cast<T*>(&storage_);
-        }
-        T&& operator*() && noexcept
-        {
-            return HPX_MOVE(*reinterpret_cast<T*>(&storage_));
-        }
-        T const&& operator*() const&& noexcept
-        {
-            return HPX_MOVE(*reinterpret_cast<T*>(&storage_));
-        }
-
-        constexpr explicit operator bool() const noexcept
-        {
-            return !empty_;
-        }
-
-        constexpr bool has_value() const
-        {
-            return !empty_;
-        }
-
-        T& value() &
-        {
-            if (empty_)
-            {
-                throw bad_optional_access(
-                    "object is empty during call to 'value()'");
-            }
-            return **this;
-        }
-        T&& value() &&
-        {
-            if (empty_)
-            {
-                throw bad_optional_access(
-                    "object is empty during call to 'value()'");
-            }
-            return **this;
-        }
-
-        constexpr T const& value() const&
-        {
-            if (empty_)
-            {
-                throw bad_optional_access(
-                    "object is empty during call to 'value()'");
-            }
-            return HPX_MOVE(**this);
-        }
-        constexpr T const&& value() const&&
-        {
-            if (empty_)
-            {
-                throw bad_optional_access(
-                    "object is empty during call to 'value()'");
-            }
-            return HPX_MOVE(**this);
-        }
-
-        template <typename U>
-        constexpr T value_or(U&& value) const&
-        {
-            if (empty_)
-                return HPX_FORWARD(U, value);
-            return **this;
-        }
-        template <typename U>
-        T value_or(U&& value) &&
-        {
-            if (empty_)
-                return HPX_FORWARD(U, value);
-            return HPX_MOVE(**this);
-        }
-
-        template <typename... Ts>
-        void emplace(Ts&&... ts)
-        {
-            if (!empty_)
-            {
-                reinterpret_cast<T*>(&storage_)->~T();
-                empty_ = true;
-            }
-            new (&storage_) T(HPX_FORWARD(Ts, ts)...);
-            empty_ = false;
-        }
-
-#if !defined(HPX_HAVE_CXX17_COPY_ELISION)
-        // workaround for broken return type copy elison in MSVC
-        template <typename F, typename... Ts>
-        void emplace_f(F&& f, Ts&&... ts)
-        {
-            if (!empty_)
-            {
-                reinterpret_cast<T*>(&storage_)->~T();
-                empty_ = true;
-            }
-            new (&storage_) T(HPX_FORWARD(F, f)(HPX_FORWARD(Ts, ts)...));
-            empty_ = false;
-        }
-#endif
-
-        void swap(optional& other) noexcept(
-            std::is_nothrow_move_constructible_v<T>&& noexcept(
-                _optional_swap::check_swap<T>()))
-        {
-            // do nothing if both are empty
-            if (empty_ && other.empty_)
-            {
-                return;
-            }
-
-            // swap content if both are non-empty
-            if (!empty_ && !other.empty_)
-            {
-                using std::swap;
-                swap(**this, *other);
-                return;
-            }
-
-            // move the non-empty one into the empty one and make remains empty
-            optional* empty = empty_ ? this : &other;
-            optional* non_empty = empty_ ? &other : this;
-
-            new (&empty->storage_) T(HPX_MOVE(**non_empty));
-            reinterpret_cast<T*>(&non_empty->storage_)->~T();
-
-            empty->empty_ = false;
-            non_empty->empty_ = true;
-        }
-
-        void reset() noexcept
-        {
-            if (!empty_)
-            {
-                reinterpret_cast<T*>(&storage_)->~T();
-                empty_ = true;
-            }
-        }
-
-        template <typename F>
-        auto and_then(F&& f) &
-        {
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return std::invoke(HPX_FORWARD(F, f), value());
-            }
-
-            using result_type = std::invoke_result_t<F, decltype(value())>;
-            return std::remove_cv_t<std::remove_reference_t<result_type>>();
-        }
-        template <typename F>
-        constexpr auto and_then(F&& f) const&
-        {
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return std::invoke(HPX_FORWARD(F, f), value());
-            }
-
-            using result_type = std::invoke_result_t<F, decltype(value())>;
-            return std::remove_cv_t<std::remove_reference_t<result_type>>();
-        }
-        template <typename F>
-        auto and_then(F&& f) &&
-        {
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return std::invoke(HPX_FORWARD(F, f), HPX_MOVE(value()));
-            }
-
-            using result_type =
-                std::invoke_result_t<F, decltype(HPX_MOVE(value()))>;
-            return std::remove_cv_t<std::remove_reference_t<result_type>>();
-        }
-        template <typename F>
-        auto and_then(F&& f) const&&
-        {
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return std::invoke(HPX_FORWARD(F, f), HPX_MOVE(value()));
-            }
-
-            using result_type =
-                std::invoke_result_t<F, decltype(HPX_MOVE(value()))>;
-            return std::remove_cv_t<std::remove_reference_t<result_type>>();
-        }
-
-        template <typename F>
-        auto transform(F&& f) &
-        {
-            using result_type = std::invoke_result_t<F, decltype(value())>;
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return optional<result_type>(
-                    std::invoke(HPX_FORWARD(F, f), value()));
-            }
-            return optional<result_type>();
-        }
-        template <typename F>
-        constexpr auto transform(F&& f) const&
-        {
-            using result_type = std::invoke_result_t<F, decltype(value())>;
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return optional<result_type>(
-                    std::invoke(HPX_FORWARD(F, f), value()));
-            }
-            return optional<result_type>();
-        }
-        template <typename F>
-        auto transform(F&& f) &&
-        {
-            using result_type =
-                std::invoke_result_t<F, decltype(HPX_MOVE(value()))>;
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return optional<result_type>(
-                    std::invoke(HPX_FORWARD(F, f), HPX_MOVE(value())));
-            }
-            return optional<result_type>();
-        }
-        template <typename F>
-        auto transform(F&& f) const&&
-        {
-            using result_type =
-                std::invoke_result_t<F, decltype(HPX_MOVE(value()))>;
-            if (*this)
-            {
-                // using std::invoke to avoid circular dependencies
-                return optional<result_type>(
-                    std::invoke(HPX_FORWARD(F, f), HPX_MOVE(value())));
-            }
-            return optional<result_type>();
-        }
-
-        template <typename F,
-            typename = std::enable_if_t<std::is_invocable_v<F> &&
-                std::is_copy_constructible_v<T>>>
-        constexpr optional or_else(F&& f) const&
-        {
-            if (*this)
-            {
-                return *this;
-            }
-            return {HPX_FORWARD(F, f)()};
-        }
-        template <typename F,
-            typename = std::enable_if_t<std::is_invocable_v<F> &&
-                std::is_move_constructible_v<T>>>
-        optional or_else(F&& f) &&
-        {
-            if (*this)
-            {
-                return HPX_MOVE(*this);
-            }
-            return {HPX_FORWARD(F, f)()};
-        }
-
-    private:
-        typename std::aligned_storage_t<sizeof(T), alignof(T)> storage_;
-        bool empty_;
-    };
-
-    ///////////////////////////////////////////////////////////////////////////
-    template <typename T>
-    constexpr bool operator==(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return (bool(lhs) != bool(rhs)) ? false :
-            (!bool(lhs) && !bool(rhs))  ? true :
-                                          (*lhs == *rhs);
-    }
-
-    template <typename T>
-    constexpr bool operator!=(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return !(lhs == rhs);
-    }
-
-    template <typename T>
-    constexpr bool operator<(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return (!bool(rhs)) ? false : (!bool(lhs)) ? true : *rhs < *lhs;
-    }
-
-    template <typename T>
-    constexpr bool operator>=(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return !(lhs < rhs);
-    }
-
-    template <typename T>
-    constexpr bool operator>(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return (!bool(lhs)) ? false : (!bool(rhs)) ? true : *rhs > *lhs;
-    }
-
-    template <typename T>
-    constexpr bool operator<=(optional<T> const& lhs, optional<T> const& rhs)
-    {
-        return !(lhs > rhs);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    template <typename T>
-    constexpr bool operator==(optional<T> const& opt, nullopt_t) noexcept
-    {
-        return !bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator==(nullopt_t, optional<T> const& opt) noexcept
-    {
-        return !bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator!=(optional<T> const& opt, nullopt_t) noexcept
-    {
-        return bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator!=(nullopt_t, optional<T> const& opt) noexcept
-    {
-        return bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator<(optional<T> const&, nullopt_t) noexcept
-    {
-        return false;
-    }
-
-    template <typename T>
-    constexpr bool operator<(nullopt_t, optional<T> const& opt) noexcept
-    {
-        return bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator>=(optional<T> const&, nullopt_t) noexcept
-    {
-        return true;
-    }
-
-    template <typename T>
-    constexpr bool operator>=(nullopt_t, optional<T> const& opt) noexcept
-    {
-        return !bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator>(optional<T> const& opt, nullopt_t) noexcept
-    {
-        return bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator>(nullopt_t, optional<T> const&) noexcept
-    {
-        return false;
-    }
-
-    template <typename T>
-    constexpr bool operator<=(optional<T> const& opt, nullopt_t) noexcept
-    {
-        return !bool(opt);
-    }
-
-    template <typename T>
-    constexpr bool operator<=(nullopt_t, optional<T> const&) noexcept
-    {
-        return true;
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    template <typename T>
-    constexpr bool operator==(optional<T> const& opt, T const& value)
-    {
-        return bool(opt) ? (*opt == value) : false;
-    }
-
-    template <typename T>
-    constexpr bool operator==(T const& value, optional<T> const& opt)
-    {
-        return bool(opt) ? (value == *opt) : false;
-    }
-
-    template <typename T>
-    constexpr bool operator!=(optional<T> const& opt, T const& value)
-    {
-        return !(opt == value);
-    }
-
-    template <typename T>
-    constexpr bool operator!=(T const& value, optional<T> const& opt)
-    {
-        return !(value == *opt);
-    }
-
-    template <typename T>
-    constexpr bool operator<(optional<T> const& opt, T const& value)
-    {
-        return bool(opt) ? (*opt < value) : true;
-    }
-
-    template <typename T>
-    constexpr bool operator<(T const& value, optional<T> const& opt)
-    {
-        return bool(opt) ? (value < *opt) : false;
-    }
-
-    template <typename T>
-    constexpr bool operator>=(optional<T> const& opt, T const& value)
-    {
-        return !(*opt < value);
-    }
-
-    template <typename T>
-    constexpr bool operator>=(T const& value, optional<T> const& opt)
-    {
-        return !(value < *opt);
-    }
-
-    template <typename T>
-    constexpr bool operator>(optional<T> const& opt, T const& value)
-    {
-        return bool(opt) ? (*opt > value) : false;
-    }
-
-    template <typename T>
-    constexpr bool operator>(T const& value, optional<T> const& opt)
-    {
-        return bool(opt) ? (value > *opt) : true;
-    }
-
-    template <typename T>
-    constexpr bool operator<=(optional<T> const& opt, T const& value)
-    {
-        return !(*opt > value);
-    }
-
-    template <typename T>
-    constexpr bool operator<=(T const& value, optional<T> const& opt)
-    {
-        return !(value > *opt);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    template <typename T>
-    void swap(optional<T>& x, optional<T>& y) noexcept(noexcept(x.swap(y)))
-    {
-        x.swap(y);
-    }
+    using std::bad_optional_access;
+    using std::make_optional;
+    using std::nullopt;
+    using std::nullopt_t;
+    using std::optional;
 }    // namespace hpx
+
+#else
+
+#include <hpx/datastructures/detail/optional.hpp>
+
+namespace hpx {
+
+    using hpx::optional_ns::bad_optional_access;
+    using hpx::optional_ns::make_optional;
+    using hpx::optional_ns::nullopt;
+    using hpx::optional_ns::nullopt_t;
+    using hpx::optional_ns::optional;
+}    // namespace hpx
+
+#endif
 
 namespace hpx::util {
 
@@ -712,7 +44,7 @@ namespace hpx::util {
 
     HPX_DEPRECATED_V(1, 8,
         "hpx::util::nullopt is deprecated. Please use hpx::nullopt instead.")
-    constexpr hpx::nullopt_t nullopt{hpx::nullopt_t::init()};
+    constexpr hpx::nullopt_t nullopt = hpx::nullopt;
 
     template <typename T>
     using optional HPX_DEPRECATED_V(1, 8,
@@ -728,39 +60,26 @@ namespace hpx::util {
     HPX_DEPRECATED_V(1, 8,
         "hpx::util::make_optional is deprecated. Please use hpx::optional "
         "instead.")
-    constexpr hpx::optional<std::decay_t<T>> make_optional(T&& v)
+    constexpr auto make_optional(T&& t)
     {
-        return hpx::optional<std::decay_t<T>>(HPX_FORWARD(T, v));
+        return hpx::make_optional(HPX_FORWARD(T, t));
     }
 
     template <typename T, typename... Ts>
     HPX_DEPRECATED_V(1, 8,
         "hpx::util::make_optional is deprecated. Please use hpx::optional "
         "instead.")
-    constexpr hpx::optional<T> make_optional(Ts&&... ts)
+    constexpr auto make_optional(Ts&&... ts)
     {
-        return hpx::optional<T>(std::in_place, HPX_FORWARD(Ts, ts)...);
+        return hpx::make_optional(HPX_FORWARD(Ts, ts)...);
     }
 
     template <typename T, typename U, typename... Ts>
     HPX_DEPRECATED_V(1, 8,
         "hpx::util::make_optional is deprecated. Please use hpx::optional "
         "instead.")
-    constexpr hpx::optional<T> make_optional(
-        std::initializer_list<U> il, Ts&&... ts)
+    constexpr auto make_optional(std::initializer_list<U> il, Ts&&... ts)
     {
-        return hpx::optional<T>(std::in_place, il, HPX_FORWARD(Ts, ts)...);
+        return hpx::make_optional(il, HPX_FORWARD(Ts, ts)...);
     }
 }    // namespace hpx::util
-
-///////////////////////////////////////////////////////////////////////////////
-namespace std {
-    template <typename T>
-    struct hash<hpx::optional<T>>
-    {
-        constexpr std::size_t operator()(::hpx::optional<T> const& arg) const
-        {
-            return arg ? std::hash<T>{}(*arg) : std::size_t{};
-        }
-    };
-}    // namespace std
