+++ libs/core/execution/include/hpx/execution/algorithms/when_all.hpp
@@ -92,12 +92,14 @@ namespace hpx::execution::experimental {
 
             template <typename... Ts, std::size_t... Is>
             auto set_value_helper(hpx::util::index_pack<Is...>, Ts&&... ts)
-                -> decltype((
-                    std::declval<
-                        typename OperationState::value_types_storage_type>()
-                        .template get<OperationState::i_storage_offset + Is>()
-                        .emplace(HPX_FORWARD(Ts, ts)),
-                    ...))
+                -> decltype(
+                    (std::declval<
+                         typename OperationState::value_types_storage_type>()
+                            .template get<OperationState::i_storage_offset +
+                                Is>()
+                            .emplace(HPX_FORWARD(Ts, ts)),
+                        ...),
+                    void())
             {
                 // op_state.ts holds values from all predecessor senders. We
                 // emplace the values using the offset calculated while
@@ -161,7 +163,7 @@ namespace hpx::execution::experimental {
         template <typename OperationState, typename... Ts>
         auto tag_invoke(set_value_t, when_all_receiver<OperationState>&& r,
             Ts&&... ts) noexcept
-            -> decltype(r.set_value(HPX_FORWARD(Ts, ts)...))
+            -> decltype(r.set_value(HPX_FORWARD(Ts, ts)...), void())
         {
             r.set_value(HPX_FORWARD(Ts, ts)...);
         }
